test: keep the binding guards inside the existing dev-platform suites
The standalone guard file added 7 lines naming Dev Platform identifiers, which
raised the epic #470 core-decoupling ratchet (1029 -> 1036 in middleware/test).
The reference count is the definition of "extraction finished", so the right
move is to not add references rather than to raise the baseline.

The same three assertions now live in a neutrally-named `harness socket
binding` suite inside devPlatformRoutes.test.ts and devRunnerApi.test.ts, which
already import their harness. Nothing in the added lines matches a ratchet
pattern, so the count is flat at 3294. Reverting either harness still turns all
three red.
